better type

+++ nucleo/cLcd.cpp
@@ -592,7 +592,7 @@ void cLcd::rgb888to565 (uint8_t* src, uint16_t* dst, uint16_t xsize) {
   }
 //}}}
 //{{{
-void cLcd::jpegYuvTo565 (uint8_t* src, uint32_t* dst, uint16_t xsize, uint16_t ysize, uint32_t chromaSampling) {
+void cLcd::jpegYuvTo565 (uint8_t* src, uint16_t* dst, uint16_t xsize, uint16_t ysize, uint32_t chromaSampling) {
 
   ready();
